+++ src/main/java/com/nylas/delta/DeltaStreamListener.java
@@ -1,0 +1,8 @@
+package com.nylas.delta;
+
+import com.nylas.AccountOwnedModel;
+
+public interface DeltaStreamListener {
+
+	void onDelta(Delta<? extends AccountOwnedModel> delta);
+}